requirements.txt
# ── Core CV & scientific ──────────────────────────────────────
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0
scikit-image>=0.21.0

# ── Deep Learning ─────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# ── Augmentation ─────────────────────────────────────────────
albumentations>=1.3.0
Pillow>=10.0.0

# ── Data & analysis ───────────────────────────────────────────
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# ── API deployment ────────────────────────────────────────────
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.3.0

# ── Experiment tracking (optional) ────────────────────────────
mlflow>=2.7.0

# ── Utilities ─────────────────────────────────────────────────
pyyaml>=6.0
loguru>=0.7.0
click>=8.1.0
python-dotenv>=1.0.0

# ── Testing ───────────────────────────────────────────────────
pytest>=7.4.0
httpx>=0.25.0
